Added and tested all functionality, Getting Orders/Products now has filtering, pagination and sorting, Products in use wont get deleted, Delivered Orders cannot be deleted

+++ Application/Order/Handlers/DeleteOrderCommandHandler.cs
@@ -23,6 +23,11 @@ namespace Application.Handlers
             {
                 return -1;
             }
+            if(order.DeliveryTime < DateTime.Now)
+            {
+
+               return -2;
+            }
             await _uow.OrderRepo.Delete(request.Id);
             return await _uow.SaveChangesAsync(cancellationToken);
         }
